pyproject.toml
[tool.poetry]
name = "book-exchange-app"
version = "0.1.0"
description = "An open-source FastAPI application for school families to exchange textbooks and reading books."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "book_exchange_app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
sqlmodel = "^0.0.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
jinja2 = "^3.1.2"        # if you use templates
python-dotenv = "^1.0.0" # for loading .env in Docker

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
httpx = "^0.28.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"

[tool.poetry.scripts]
start = "uvicorn book_exchange_app.main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"
